Move Signup page to Signup feature folder
