+++ Elecritic/Features/Signup/Pages/SignupPage.razor
@@ -1,0 +1,53 @@
+﻿@page "/signup"
+
+<AuthorizeView>
+    <Authorized>
+        <p><b>Actualmente ya tienes una sesión activa, con correo.</b></p>
+    </Authorized>
+    <NotAuthorized Context="notAuthContext">
+        <div class="middle-box">
+            <h3>Registro</h3>
+            <div class="row">
+                <div class="col-12">
+                    <h2>Crea una nueva cuenta</h2>
+                    <EditForm Model="@SignupForm" OnValidSubmit="async () => await SignUpAsync()">
+                        <RadzenCard>
+                            @*<AlertMessage Message="@message" MessageType="@messageType" />*@
+                            <DataAnnotationsValidator />
+                            <ValidationSummary />
+
+                            <label>Nombre de usuario</label>
+                            <RadzenTextBox @bind-Value="@SignupForm.Username"></RadzenTextBox>
+
+                            <label>Primer nombre</label>
+                            <RadzenTextBox @bind-Value="@SignupForm.FirstName"></RadzenTextBox>
+
+                            <label>Apellido</label>
+                            <RadzenTextBox @bind-Value="@SignupForm.LastName"></RadzenTextBox>
+
+                            <label>Correo electrónico</label>
+                            <RadzenTextBox @bind-Value="@SignupForm.Email"></RadzenTextBox>
+
+                            <label>Contraseña</label>
+                            <RadzenPassword @bind-Value="@SignupForm.Password"></RadzenPassword>
+
+                            <label>Confirmar contraseña</label>
+                            <RadzenPassword @bind-Value="@SignupForm.ConfirmPassword" class="finalBox"></RadzenPassword>
+
+                            <RadzenButton ButtonType="@ButtonType.Submit" Text="Crear cuenta"
+                                          ButtonStyle="@ButtonStyle.Primary" class="btn-block mr-2"
+                                          Disabled="@IsSigningUp" />
+                            <hr />
+                            <label class="labelCenter">¿Ya tienes una cuenta?</label>
+                            <RadzenButton ButtonType="@ButtonType.Button" Text="Inicia sesión"
+                                          ButtonStyle="@ButtonStyle.Secondary" class="btn-block mr-2"
+                                          Disabled="@IsSigningUp" Click="@GoToLogin" />
+                        </RadzenCard>
+                    </EditForm>
+                </div>
+            </div>
+        </div>
+
+        <p>@ResultMessage</p>
+    </NotAuthorized>
+</AuthorizeView>
